feat: add main header and fixed homeHeader
