+++ Frontend/src/components/headPack/appHeader/appheader.module.css
@@ -1,0 +1,13 @@
+.header {
+    width: 100%;
+    height: 110px;
+    background: linear-gradient(to right, #1E90FF 67%, #1E90FF 100%);
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 20px 50px;
+}
+
+.logo {
+    padding-top: 18px;
+}
