--- Frontend/src/components/headPack/header/header.module.css
@@ -1,9 +1,0 @@
-.header {
-    width: 100%;
-    height: 124px;
-    background-color: white;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 20px 50px; /*переделать и настроить соотношение расположения в шапке */
-}
